requirements.txt
slack-bolt
python-dotenv
flask
supabase
requests 